Support fractional scale for `Sprite`.
Although the majority of cases involve scaling up, where `int` is fine, there may be a few cases where images need to be scaled down, especially when using external (mod-added) higher-res images.

If these images are nine-patched, then the entire sprite should be interpreted at smaller scale so that it can be scaled back up again in terms of layout, but still maintaining the smaller border slices.

+++ src/NineSlice.cs
@@ -100,7 +100,7 @@ public class NineSlice(Sprite sprite)
         SliceSettings? settings = null,
         // We pass sliceScale separately (even though it is in SliceSettings) because it actually applies only to the
         // destination rect, not the source, whereas the SliceSettings are used for source but not destination.
-        int sliceScale = 1
+        float sliceScale = 1
     )
     {
         var left = bounds.X;
